requirements.txt
flask
google-cloud-firestore
google-cloud-aiplatform